backend/requirements.txt
# Python dependencies for Clinic AI Bot
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
groq>=0.4.0
python-multipart>=0.0.6
APScheduler>=3.10.0
gunicorn>=21.0.0
